Modify column names to bigquery naming requirements
When column names do not start with a letter or underscore
Google BigQuery will prevent creation of the column and the
loading of the data will fail. Columns that have other characters
as their first letter will have the 'col_' prefix added.

Additionally, any columns that are greater than or equal to 40
characters will have this prefix appended to them. This is to
ensure uniformity with Pipedrive data.